Remove `HostLowLevelDevice` header
