+++ src/runtime/micro/host_low_level_device.cc
@@ -26,41 +26,68 @@
 #include <sys/mman.h>
 #include <cstring>
 #include <memory>
-#include "host_low_level_device.h"
 #include "micro_common.h"
+#include "low_level_device.h"
 
 namespace tvm {
 namespace runtime {
 
-HostLowLevelDevice::HostLowLevelDevice(size_t num_bytes) : size_(num_bytes) {
-  size_t size_in_pages = (num_bytes + kPageSize - 1) / kPageSize;
-  // TODO(weberlo): Set permissions per section (e.g., read-write perms for
-  // the heap, execute perms for text, etc.).
-  int mmap_prot = PROT_READ | PROT_WRITE | PROT_EXEC;
-  int mmap_flags = MAP_ANONYMOUS | MAP_PRIVATE;
-  base_addr_ = DevBaseAddr(
-    (reinterpret_cast<std::uintptr_t>(
-      mmap(nullptr, size_in_pages * kPageSize, mmap_prot, mmap_flags, -1, 0))));
-}
+/*!
+ * \brief emulated low-level device on host machine
+ */
+class HostLowLevelDevice final : public LowLevelDevice {
+ public:
+  /*!
+   * \brief constructor to initialize on-host memory region to act as device
+   * \param num_bytes size of the emulated on-device memory region
+   */
+  explicit HostLowLevelDevice(size_t num_bytes) : size_(num_bytes) {
+    size_t size_in_pages = (num_bytes + kPageSize - 1) / kPageSize;
+    // TODO(weberlo): Set permissions per section (e.g., read-write perms for
+    // the heap, execute perms for text, etc.).
+    int mmap_prot = PROT_READ | PROT_WRITE | PROT_EXEC;
+    int mmap_flags = MAP_ANONYMOUS | MAP_PRIVATE;
+    base_addr_ = DevBaseAddr(
+      (reinterpret_cast<std::uintptr_t>(
+        mmap(nullptr, size_in_pages * kPageSize, mmap_prot, mmap_flags, -1, 0))));
+  }
 
-HostLowLevelDevice::~HostLowLevelDevice() {
-  munmap(base_addr_.cast_to<void*>(), size_);
-}
+  /*!
+   * \brief destructor to deallocate on-host device region
+   */
+  virtual ~HostLowLevelDevice() {
+    munmap(base_addr_.cast_to<void*>(), size_);
+  }
 
-void HostLowLevelDevice::Read(DevBaseOffset offset, void* buf, size_t num_bytes) {
-  void* addr = (offset + base_addr_).cast_to<void*>();
-  std::memcpy(buf, addr, num_bytes);
-}
+  void Read(DevBaseOffset offset, void* buf, size_t num_bytes) {
+    void* addr = (offset + base_addr_).cast_to<void*>();
+    std::memcpy(buf, addr, num_bytes);
+  }
 
-void HostLowLevelDevice::Write(DevBaseOffset offset, const void* buf, size_t num_bytes) {
-  void* addr = (offset + base_addr_).cast_to<void*>();
-  std::memcpy(addr, buf, num_bytes);
-}
+  void Write(DevBaseOffset offset, const void* buf, size_t num_bytes) {
+    void* addr = (offset + base_addr_).cast_to<void*>();
+    std::memcpy(addr, buf, num_bytes);
+  }
 
-void HostLowLevelDevice::Execute(DevBaseOffset func_offset, DevBaseOffset breakpoint) {
-  DevAddr func_addr = func_offset + base_addr_;
-  reinterpret_cast<void (*)(void)>(func_addr.value())();
-}
+  void Execute(DevBaseOffset func_offset, DevBaseOffset breakpoint) {
+    DevAddr func_addr = func_offset + base_addr_;
+    reinterpret_cast<void (*)(void)>(func_addr.value())();
+  }
+
+  DevBaseAddr base_addr() const final {
+    return base_addr_;
+  }
+
+  const char* device_type() const final {
+    return "host";
+  }
+
+ private:
+  /*! \brief base address of the micro device memory region */
+  DevBaseAddr base_addr_;
+  /*! \brief size of memory region */
+  size_t size_;
+};
 
 const std::shared_ptr<LowLevelDevice> HostLowLevelDeviceCreate(size_t num_bytes) {
   std::shared_ptr<LowLevelDevice> lld =
